+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+- Fixed a bug where stores’ “Order Reference Number Format” weren’t being saved to the database. ([#3454](https://github.com/craftcms/commerce/issues/3454))
+
 ## 5.0.0-beta.3 - 2024-04-10
 
 - Removed the `showEditUserCommerceTab` config setting.